Cambios Version 2

+++ resources/views/congresos/index.blade.php
@@ -3,12 +3,15 @@
 	{!! Breadcrumbs::render('congresos') !!}
 			<div class="panel panel-default">
 				<div class="panel-body">
-					<div class="col-lg-12 margin-tb">
+					<div class="col-sm-12 margin-tb">
 						<div class="pull-left">
 							<h2>{{ __('Kongresu Zientifikoetan parte-hartzea') }}</h2>
 						</div>
+						<div class="pull-left margen-left">
+							<a class="btn btn-info" href="{{ route('congresos.indexAll') }}"><i class="fa fa-eye" title="{{ __('Kongresu zientifiko guztiak ikusi') }}"></i></a>
+						</div>
 						<div class="pull-right">
-							<a class="btn btn-success" href="{{ route('congresos.create') }}"> {{ __('Berria sortu') }}</a>
+							<a class="btn btn-success" href="{{ route('congresos.create') }}"><i class="fa fa-plus" title ="{{ __('Berria sortu') }}"></i></a>
 						</div>
 					</div>
 				</div>
@@ -19,15 +22,19 @@
 				@endif
 				<table class="table">
 					<tr>
-						<th>Kongresu</th>
-						<th>Congreso</th>
-						<th>{{ __('Arduraduna') }}</th>
+						<th>{{ __('Kongresu') }}</th>
+						<th>{{ __('Arduraduna(k)') }}</th>
 						<th>{{ __('Akzioak') }}</th>
 					</tr>
 					@foreach ($data as $key => $congreso)
 					<tr>
-						<td>{{ $congreso->congreso_eu }}</td>
-						<td>{{ $congreso->congreso_es }}</td>
+						<td>
+							<?php $congre = "congreso_".\Session::get('locale') ;?>
+								<a href="{{ route('congresos.edit',$congreso->id) }}">
+									{{ $congreso->$congre }}, ( {{ $congreso->desde }} - {{ $congreso->hasta }} )
+									</a>
+							<br> <i>({{ $congreso->usuario?$congreso->usuario->name:'' }} {{ $congreso->usuario?$congreso->usuario->lname:'' }})</i>
+						</td>
 						<td>
 								@foreach( $congreso->profesores as $profesor)
 			 					{{$profesor->nombre}} {{$profesor->apellido}}
@@ -37,14 +44,16 @@
 			 				@endforeach
 						</td>
 						<td>
-							<a class="btn btn-primary" href="{{ route('congresos.edit',$congreso->id) }}">{{ __('Aldatu')}}</a>
-							{!! Form::open(['method' => 'DELETE','route' => ['congresos.destroy', $congreso->id],'style'=>'display:inline']) !!}
-							{!! Form::submit(__('Ezabatu'), ['class' => 'btn btn-danger']) !!}
-							{!! Form::close() !!}
+							<a class="btn btn-primary" href="{{ route('congresos.edit',$congreso->id) }}"><i class="fa fa-pencil" title="{{ __('Aldadtu') }}"></i></a>
+							@if( $congreso->user_id == \Auth::user()->id )
+								{!! Form::open(['method' => 'DELETE','route' => ['congresos.destroy', $congreso->id],'style'=>'display:inline']) !!}
+								{{ Form::button('<i class="fa fa-trash"  title="'.__('Ezabatu').'"></i> ', ['type' => 'submit', 'class' => 'btn btn-danger'] )  }}
+								{!! Form::close() !!}
+							@endif
 						</td>
 					</tr>
 					@endforeach
-					<tr><td>{{ __('Guztira:' )}} {{ $data->total() }}</td><td colspan='2' class='text-center'>{{ $data->links() }}</td><td>{{ __('Pagina actual:' )}} {{ $data->currentPage() }}</td></tr>
+					<tr><td>{{ __('Guztira:' )}} {{ $data->total() }}</td><td  class='text-center'>{{ $data->links() }}</td><td>{{ __('Oraingo orria:' )}} {{ $data->currentPage() }}</td></tr>
 				</table>
 			</div>
 @endsection
