+++ resources/views/publicaciones/index.blade.php
@@ -3,7 +3,7 @@
 	{!! Breadcrumbs::render('publicaciones', $tipo) !!}
 			<div class="panel panel-default">
 				<div class="panel-body">
-					<div class="col-lg-12 margin-tb">
+					<div class="col-sm-12 margin-tb">
 						<div class="pull-left">
 							@if( $tipo == 'libros' )
 								<h2>{{ __('Liburuak eta Monografiak') }}</h2>
@@ -11,8 +11,11 @@
     							<h2>{{ __('Artikuloak') }}</h2>
 							@endif
 						</div>
+						<div class="pull-left margen-left">
+							<a class="btn btn-info" href="{{ route('publicaciones.indexAll', [ 'tipo'=> $tipo ]) }}"><i class="fa fa-eye" title="{{ __('Guztiak ikusi') }}"></i></a>
+						</div>
 						<div class="pull-right">
-							<a class="btn btn-success" href="{{ route('publicaciones.create', [ 'tipo'=> $tipo ] ) }}"> {{ __('Berria sortu') }}</a>
+							<a class="btn btn-success" href="{{ route('publicaciones.create', [ 'tipo'=> $tipo ] ) }}"><i class="fa fa-plus" title ="{{ __('Berria sortu') }}"></i></a>
 						</div>
 					</div>
 				</div>
@@ -24,15 +27,17 @@
 				@endif
 				<table class="table">
 					<tr>
-						<th>Izenburua</th>
-						<th>Titulo</th>
-						<th>{{ __('Egileak') }}</th>
+						<th>{{ __('Izenburua') }}</th>
+						<th>{{ __('Egilea(k)') }}</th>
 						<th>{{ __('Akzioak') }}</th>
 					</tr>
 					@foreach ($data as $key => $publicacion)
 					<tr>
-						<td>{{ $publicacion->titulo_eu }}</td>
-						<td>{{ $publicacion->titulo_es }}</td>
+						<td>
+							<?php $titulo = "titulo_".\Session::get('locale') ;?>
+							<a  href="{{ route('publicaciones.edit',$publicacion->id) }}">{{ $publicacion->$titulo }}</a>
+							<br> <i>({{ $publicacion->usuario?$publicacion->usuario->name:'' }} {{ $publicacion->usuario?$publicacion->usuario->lname:'' }})</i>
+						</td>
 						<td>
 							@foreach( $publicacion->autores as $autor)
 			 					{{$autor->nombre}} {{$autor->apellido}}
@@ -42,14 +47,16 @@
 			 				@endforeach
 						</td>
 						<td>
-							<a class="btn btn-primary" href="{{ route('publicaciones.edit',$publicacion->id) }}">{{ __('Aldatu')}}</a>
-							{!! Form::open(['method' => 'DELETE','route' => ['publicaciones.destroy', $publicacion->id, $publicacion->tipo],'style'=>'display:inline']) !!}
-							{!! Form::submit(__('Ezabatu'), ['class' => 'btn btn-danger']) !!}
-							{!! Form::close() !!}
+							<a class="btn btn-primary" href="{{ route('publicaciones.edit',$publicacion->id) }}"><i class="fa fa-pencil" title="{{ __('Aldadtu') }}"></i></a>
+							@if( $publicacion->user_id == \Auth::user()->id )
+								{!! Form::open(['method' => 'DELETE','route' => ['publicaciones.destroy', $publicacion->id, $publicacion->tipo],'style'=>'display:inline']) !!}
+								{{ Form::button('<i class="fa fa-trash"  title="'.__('Ezabatu').'"></i> ', ['type' => 'submit', 'class' => 'btn btn-danger'] )  }}
+								{!! Form::close() !!}
+							@endif
 						</td>
 					</tr>
 					@endforeach
-					<tr><td>{{ __('Guztira:' )}} {{ $data->total() }}</td><td colspan='2' class='text-center'>{{ $data->links() }}</td><td>{{ __('Pagina actual:' )}} {{ $data->currentPage() }}</td></tr>
+					<tr><td>{{ __('Guztira:' )}} {{ $data->total() }}</td><td class='text-center'>{{ $data->links() }}</td><td>{{ __('Oraingo orria:' )}} {{ $data->currentPage() }}</td></tr>
 				</table>
 			</div>
 @endsection
